+++ docs/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Job types can also retrieve that information by extracting it from an HTTP header.
   [issue #246](https://github.com/TheRacetrack/racetrack/issues/246)
 
+### Changed
+- racetrack-client shows more details in case of an HTTP error.
+  [issue #245](https://github.com/TheRacetrack/racetrack/issues/245)
+
 ## [2.14.0] - 2023-05-18
 ### Added
 - Dashboard links to job-related Grafana dashobards.
@@ -29,8 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Panel takes up the whole space available - no white bars on the sides
   - Jobs tree has its own scrollbar
   - "10 seconds ago" labels are refreshed over time.
-- racetrack-client shows more details in case of an HTTP error.
-  [issue #245](https://github.com/TheRacetrack/racetrack/issues/245)
 
 ### Fixed
 - Tree of jobs is refreshed after deleting a job.